fix: nav bar divider background was dropped by SwipeBackLayout
The navbar background height is lower than the navbar height on some devices, such that the height difference area is the navbar divider consists of part of the window background. Thus we can not just set the window background to be transparent, in case an abruptly colored divider occurs there.